kite_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kiteconnect==4.2.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.24.4
ta==0.10.2
google-generativeai==0.3.2